// EmergencyBurial/Website/src/app/shared/enum/deceased-process-status.enum.ts
export enum DeceasedProcessStatus {

  PoliceIntake = 1,

  ReceptionAtTarah = 2,

  ReleaseFromTarah = 3,

  TransportToBurialPreparation = 4,

  EndTransportBurialPreparation = 5,

  ReceivedForBurialPreparation = 6,

  ReleasedFromBurialPreparation = 7,

  TransportToBurialEntity = 8,

  EndTransportBurialEntity = 9,

  Burial = 10,

  Archive = 11
}